refactor(editor): share selection handles + resize/rotate math (text ↔ frame)
The frame overlay had its own copy of the selection chrome + interaction math.
Extract them so the text tool and the frame tool use ONE implementation:

- SelectionHandles.jsx — the dashed box + 8 resize handles + rotate handle,
  lifted verbatim from TextCanvas (props: onResizeStart / onRotateStart).
- selectionMath.js — snapAngle (cardinal rotation snap) + resizeRatio
  (distance-from-center), verbatim-equivalent to TextCanvas's inline math.

TextCanvas and FrameEditOverlay both consume them; TextCanvas's local
SelectionOverlay + inline snap/ratio removed. Behavior unchanged — golden e2e
03/06/18/19/20 green.
